🚧 Add event handler for transfer and purchase (#884)
